paras: split tests
This is only a module structure change: the tests module was promoted to
have its own file.
